Add `enabled-features` fn
Some parts of the system may not have live access to redis (eg. a
Clojurescript front end). This commit provides an easy way to get the
set of all features for a given akey as data (versus checking them
individually with `enabled?` ... the set itself can act as that
predicate via a simple membership check).